+++ app/sprinkles/core/tests/Unit/BelongsToManyThroughTest.php
@@ -7,20 +7,15 @@
  */
 namespace UserFrosting\Tests\Unit;
 
-use Illuminate\Database\Capsule\Manager as DB;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Mockery as m;
-use ReflectionClass;
 use UserFrosting\Tests\TestCase;
-use UserFrosting\Tests\DatabaseTransactions;
 use UserFrosting\Sprinkle\Core\Database\Builder as QueryBuilder;
 use UserFrosting\Sprinkle\Core\Database\Models\Model;
 use UserFrosting\Sprinkle\Core\Database\Relations\BelongsToManyThrough;
 
 /**
  * Tests the BelongsToManyThrough relation.
- *
- * @extends TestCase
  */
 class BelongsToManyThroughTest extends TestCase
 {
@@ -70,7 +65,7 @@ class BelongsToManyThroughTest extends TestCase
         $related->shouldReceive('getKeyName')->andReturn('id');
         // Tie the mocked builder to the mocked related model
         $builder->shouldReceive('getModel')->andReturn($related);
-        
+
         // Mock the intermediate role->permission BelongsToMany relation
         $intermediateRelationship = m::mock(BelongsToMany::class);
         $intermediateRelationship->shouldReceive('getTable')->once()->andReturn('permission_roles');